+++ src/annotations/recentChangesAnnotationProvider.ts
@@ -1,9 +1,10 @@
 'use strict';
 import { DecorationOptions, ExtensionContext, MarkdownString, Position, Range, TextEditor, TextEditorDecorationType } from 'vscode';
-import { Annotations, endOfLineIndex } from './annotations';
 import { FileAnnotationType } from './annotationController';
 import { AnnotationProviderBase } from './annotationProvider';
-import { GitService, GitUri } from '../gitService';
+import { Annotations } from './annotations';
+import { RangeEndOfLineIndex } from '../constants';
+import { GitContextTracker, GitService, GitUri } from '../gitService';
 import { Logger } from '../logger';
 
 export class RecentChangesAnnotationProvider extends AnnotationProviderBase {
@@ -11,15 +12,16 @@ export class RecentChangesAnnotationProvider extends AnnotationProviderBase {
     constructor(
         context: ExtensionContext,
         editor: TextEditor,
+        gitContextTracker: GitContextTracker,
         decoration: TextEditorDecorationType | undefined,
         highlightDecoration: TextEditorDecorationType | undefined,
         private readonly git: GitService,
         private readonly uri: GitUri
     ) {
-        super(context, editor, decoration, highlightDecoration);
+        super(context, editor, gitContextTracker, decoration, highlightDecoration);
     }
 
-    async provideAnnotation(shaOrLine?: string | number): Promise<boolean> {
+    async onProvideAnnotation(shaOrLine?: string | number): Promise<boolean> {
         this.annotationType = FileAnnotationType.RecentChanges;
 
         const commit = await this.git.getLogCommit(this.uri.repoPath, this.uri.fsPath, { previous: true });
@@ -44,7 +46,7 @@ export class RecentChangesAnnotationProvider extends AnnotationProviderBase {
 
                 if (line.state === 'unchanged') continue;
 
-                const range = this.editor.document.validateRange(new Range(new Position(count, 0), new Position(count, endOfLineIndex)));
+                const range = this.editor.document.validateRange(new Range(new Position(count, 0), new Position(count, RangeEndOfLineIndex)));
 
                 if (cfg.hover.details) {
                     this._decorations.push({