+++ src/commands/copyMessageToClipboard.ts
@@ -52,13 +52,13 @@ export class CopyMessageToClipboardCommand extends ActiveEditorCommand {
 
             if (args.message === undefined) {
                 if (args.sha === undefined) {
-                    if (editor !== undefined && editor.document !== undefined && editor.document.isDirty) return undefined;
-
                     const blameline = (editor && editor.selection.active.line) || 0;
                     if (blameline < 0) return undefined;
 
                     try {
-                        const blame = await this.git.getBlameForLine(gitUri, blameline);
+                        const blame = editor && editor.document && editor.document.isDirty
+                            ? await this.git.getBlameForLineContents(gitUri, blameline, editor.document.getText())
+                            : await this.git.getBlameForLine(gitUri, blameline);
                         if (!blame) return undefined;
 
                         if (blame.commit.isUncommitted) return undefined;