+++ src/commands/toggleLineBlame.ts
@@ -18,7 +18,7 @@ export class ToggleLineBlameCommand extends EditorCommand {
     }
 
     async execute(editor: TextEditor, edit: TextEditorEdit, uri?: Uri, args: ToggleLineBlameCommandArgs = {}): Promise<any> {
-        if (editor === undefined || editor.document.isDirty) return undefined;
+        if (editor === undefined) return undefined;
 
         try {
             if (args.type === undefined) {