+++ src/gitCodeLensProvider.ts
@@ -69,7 +69,9 @@ export class GitCodeLensProvider implements CodeLensProvider {
     async provideCodeLenses(document: TextDocument, token: CancellationToken): Promise<CodeLens[]> {
         if (!await this.git.isTracked(document.uri.fsPath)) return [];
 
-        const dirty = document.isDirty;
+        const dirty = configuration.get<boolean>(configuration.name('insiders').value)
+            ? false
+            : document.isDirty;
 
         const cfg = configuration.get<ICodeLensConfig>(configuration.name('codeLens').value, document.uri);
         this._debug = cfg.debug;
@@ -103,7 +105,9 @@ export class GitCodeLensProvider implements CodeLensProvider {
             }
             else {
                 [blame, symbols] = await Promise.all([
-                    this.git.getBlameForFile(gitUri),
+                    document.isDirty
+                        ? this.git.getBlameForFileContents(gitUri, document.getText())
+                        : this.git.getBlameForFile(gitUri),
                     commands.executeCommand(BuiltInCommands.ExecuteDocumentSymbolProvider, document.uri) as Promise<SymbolInformation[]>
                 ]);
             }