+++ src/gitService.ts
@@ -271,7 +271,7 @@ export class GitService extends Disposable {
 
         if (!initializing) {
             // Defer the event trigger enough to let everything unwind
-            setTimeout(() => this.fireChange(GitChangeReason.Repositories), 1);
+            setImmediate(() => this.fireChange(GitChangeReason.Repositories));
         }
     }
 
@@ -687,10 +687,10 @@ export class GitService extends Disposable {
         }
     }
 
-    async getBlameForLineContents(uri: GitUri, line: number, contents: string): Promise<GitBlameLine | undefined> {
+    async getBlameForLineContents(uri: GitUri, line: number, contents: string, options: { skipCache?: boolean } = {}): Promise<GitBlameLine | undefined> {
         Logger.log(`getBlameForLineContents('${uri.repoPath}', '${uri.fsPath}', ${line})`);
 
-        if (this.UseCaching) {
+        if (!options.skipCache && this.UseCaching) {
             const blame = await this.getBlameForFileContents(uri, contents);
             if (blame === undefined) return undefined;
 
@@ -1230,11 +1230,11 @@ export class GitService extends Disposable {
             this._repositoryTree.set(rp, repo);
 
             // Send a notification that the repositories changed
-            setTimeout(async () => {
+            setImmediate(async () => {
                 await setCommandContext(CommandContext.HasRepository, this._repositoryTree.any());
 
                 this.fireChange(GitChangeReason.Repositories);
-            }, 0);
+            });
         }
 
         return rp;